Finish auth and freindships actions, add tests.

+++ Server - Clean Architecture/Use_cases/comment/delete_comment_by_id.js
@@ -1,5 +1,12 @@
-export function buildDeleteCommentById({ commentsDb }) {
+function buildDeleteCommentById({ commentsDb, Id, AppError }) {
     return async function deleteCommentById({ commentId }) {
+
+        if (!Id.isValid(commentId)) {
+            throw new AppError("Can't delete comment by invalid id.")
+        }
+
         await commentsDb.deleteById(commentId)
     }
 }
+
+module.exports = { buildDeleteCommentById }
