+++ Server - Clean Architecture/Use_cases/post/delete_post_by_id.js
@@ -1,5 +1,12 @@
-export function buildDeletePostById({ postsDb }) {
+function buildDeletePostById({ postsDb, AppError }) {
     return async function deletePostById({ postId }) {
+
+        if (!Id.isValid(postId)) {
+            throw new AppError("Can't delete post by invalid id.")
+        }
+
         await postsDb.deleteById(postId)
     }
 }
+
+module.exports = { buildDeletePostById }
