+++ Server - Clean Architecture/Use_cases/story/delete_story_by_id.js
@@ -1,5 +1,12 @@
-export function buildDeleteStoryById({ storiesDb }) {
+function buildDeleteStoryById({ storiesDb, Id, AppError }) {
     return async function deleteStoryById({ storyId }) {
+
+        if (!Id.isValid(storyId)) {
+            throw new AppError("Can't delete story by invalid id.")
+        }
+
         await storiesDb.deleteById(storyId)
     }
 }
+
+module.exports = { buildDeleteStoryById }
